Ultima Atalização

+++ src/main/java/br/unitins/dto/CidadeDTO.java
@@ -1,10 +1,10 @@
 package br.unitins.dto;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 
 public record CidadeDTO(
     @NotBlank(message = "O campo nome deve ser informado.") String nome,
-    Integer estado
+    Long estado
     ){
 }
